pyproject.toml
[project]
name = "personal-ai-agent"
version = "0.1.0"
description = "Multi-agent personal assistant using LangGraph"
requires-python = ">=3.11"
dependencies = [
    # Core LangGraph & LangChain
    "langgraph>=0.2.50",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    
    # Database for LangGraph State
    "psycopg[binary]>=3.2.0",
    
    # Google APIs (Calendar & Gmail)
    "google-api-python-client>=2.150.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    
    # Notion API
    "notion-client>=2.2.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"